requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
Flask-WTF==1.2.1
WTForms==3.1.2
wtforms-sqlalchemy==0.4.1
pandas
